various changes
redid functions so now getdev needs to be called, the n geach getframes
will return a tuple of the 3 frames and release them. finally reldev is
needed torelease device. also found memory leak. numpy needs to have
flag set to own data, this works for first data returned then throws a
memory violation when it is asked for the next tuple of frames!!!